Move non-test code to utils module; update panther_core
