+++ Pipfile
@@ -10,13 +10,13 @@ coverage = "==7.3.1"
 isort = "==5.12.0"
 mypy = "==1.5.1"
 nose2 = ">=0.13.0"
-panther-core = "==0.9.0"
+panther-core = "==0.10.0"
 pyfakefs = "==5.2.4"
 pylint = "==2.17.5"
 PyYAML = "==6.0.1"
 responses = "*"
 twine = "*"
-typed-ast = { version = "==1.4.3", markers = "python_version < '3.8' and implementation_name == 'cpython'" }
+typed-ast = "==1.5.5"
 types-python-dateutil = "*"
 types-requests = "*"
