+++ requirements.txt
@@ -37,11 +37,11 @@ pluggy==1.5.0; python_version >= '3.8'
 ply==3.11
 policyuniverse==1.5.1.20230817; python_version >= '3.7'
 pygments==2.17.2; python_version >= '3.7'
-pytest==8.1.1; python_version >= '3.8'
+pytest==8.2.0; python_version >= '3.8'
 python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
 pyyaml==6.0.1; python_version >= '3.6'
 referencing==0.35.0; python_version >= '3.8'
-regex==2024.4.16; python_version >= '3.7'
+regex==2024.4.28; python_version >= '3.8'
 requests==2.31.0; python_version >= '3.7'
 rpds-py==0.18.0; python_version >= '3.8'
 ruamel.yaml==0.17.32; python_version >= '3'
