+++ tests/unit/panther_analysis_tool/test_enriched_event_generator.py
@@ -1,7 +1,6 @@
 import copy
 import io
 import logging
-import typing
 from unittest import TestCase, mock
 from unittest.mock import call, mock_open, patch
 
@@ -12,7 +11,6 @@ from panther_analysis_tool.analysis_utils import (
 )
 from panther_analysis_tool.backend.client import (
     BackendResponse,
-    Client,
     GenerateEnrichedEventResponse,
 )
 from panther_analysis_tool.backend.mocks import MockBackend
@@ -21,71 +19,7 @@ from panther_analysis_tool.enriched_event_generator import (
     TEST_CASE_FIELD_KEY_RESOURCE,
     EnrichedEventGenerator,
 )
-
-
-def get_specs_for_test() -> typing.Dict[str, LoadAnalysisSpecsResult]:
-    return {
-        AnalysisTypes.RULE: LoadAnalysisSpecsResult(
-            f"filname.rule",
-            f"filepath.rule",
-            get_yaml_loader(roundtrip=True).load(
-                """
-                RuleID: foo.bar.rule
-                AnalysisType: rule
-                Tests:
-                  - Name: Test1
-                    ExpectedResult: true
-                    Log:
-                        a: event_type
-                        b: Equals
-                        c: 1234
-                        json: {"foo": "bar"}
-                """
-            ),
-            yaml_ctx=get_yaml_loader(roundtrip=True),
-            error=None,
-        ),
-        AnalysisTypes.SCHEDULED_RULE: LoadAnalysisSpecsResult(
-            f"filname.scheduled_rule",
-            f"filepath.scheduled_rule",
-            get_yaml_loader(roundtrip=True).load(
-                """
-                RuleID: foo.bar.scheduled_rule
-                AnalysisType: scheduled_rule
-                Tests:
-                  - Name: Test1
-                    ExpectedResult: true
-                    Log:
-                        a: event_type
-                        b: Equals
-                        c: 1234
-                        json: {"foo": "bar"}
-                """
-            ),
-            yaml_ctx=get_yaml_loader(roundtrip=True),
-            error=None,
-        ),
-        AnalysisTypes.POLICY: LoadAnalysisSpecsResult(
-            f"filname.policy",
-            f"filepath.policy",
-            get_yaml_loader(roundtrip=True).load(
-                """
-                PolicyID: foo.bar.policy
-                AnalysisType: policy
-                Tests:
-                  - Name: Test1
-                    ExpectedResult: true
-                    Resource:
-                        a: event_type
-                        b: Equals
-                        c: 1234
-                        json: {"foo": "bar"}
-                """
-            ),
-            yaml_ctx=get_yaml_loader(roundtrip=True),
-            error=None,
-        ),
-    }
+from tests.utils import get_specs_for_test
 
 
 class TestEnrichedEventGenerator(TestCase):
